Update GraphQL sample
