+++ samples/GraphQL/Models/Hero.cs
@@ -1,0 +1,7 @@
+﻿namespace Ocelot.Samples.GraphQL.Models;
+
+public class Hero
+{
+    public int Id { get; set; }
+    public required string Name { get; set; }
+}
