+++ samples/GraphQL/Ocelot.Samples.GraphQL.csproj
@@ -1,14 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>disable</ImplicitUsings>
-    <Nullable>disable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <VersionPrefix>24.0.0</VersionPrefix>
+    <AssemblyVersion>24.0.0</AssemblyVersion>
+    <Copyright>ThreeMammals</Copyright>
   </PropertyGroup>
-  <ItemGroup>
-    <None Update="ocelot.json;appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
   <ItemGroup>
     <PackageReference Include="GraphQL" Version="8.2.1" />
     <PackageReference Include="GraphQL.NewtonsoftJson" Version="8.2.1" />
